+++ src/ft2_main.c
@@ -49,6 +49,7 @@
 #include "ft2_pattern_launcher_ui.h"
 #include "ft2_splash.h"
 #include "ft2_tapesister_exchange.h"
+#include "ft2_capture.h"
 #include "ft2_palette.h"
 
 static void initializeVars(void);
@@ -322,6 +323,7 @@ int main(int argc, char *argv[])
 		readInput();
 		handleEvents();
 		tapeSisterExchangePoll(false);
+		tapeheadCapturePoll();
 		handleRecPlusExhaustion();
 		handlePolyMatrixQHandoff();
 		handlePatternLauncherStop();